+++ src/musicbrainz.c
@@ -17,6 +17,7 @@
  */
 
 #include "musicbrainz.h"
+#include "qrcode.h"
 #include "cyanrip_log.h"
 
 #include <musicbrainz5/mb5_c.h>
@@ -163,16 +164,19 @@ static int mb_tracks(cyanrip_ctx *ctx, Mb5Release release, const char *discid, i
             break;
         Mb5Track track = mb5_track_list_item(track_list, i);
         Mb5Recording recording = mb5_track_get_recording(track);
+        Mb5ArtistCredit credit = mb5_track_get_artistcredit(track);
 
         READ_MB(mb5_recording_get_id, recording, ctx->tracks[i].meta, "mbid", 0);
 
-        Mb5ArtistCredit credit;
+        /* Prefer the track title over the recording title */
+        READ_MB(mb5_track_get_title, track, ctx->tracks[i].meta, "title", 0);
+
         if (recording) {
-            READ_MB(mb5_recording_get_title, recording, ctx->tracks[i].meta, "title", 0);
-            credit = mb5_recording_get_artistcredit(recording);
-        } else {
-            READ_MB(mb5_track_get_title, track, ctx->tracks[i].meta, "title", 0);
-            credit = mb5_track_get_artistcredit(track);
+            if (!dict_get(ctx->tracks[i].meta, "title"))
+                READ_MB(mb5_recording_get_title, recording, ctx->tracks[i].meta, "title", 0);
+
+            if (!credit)
+                credit = mb5_recording_get_artistcredit(recording);
         }
         if (credit)
             mb_credit(credit, ctx->tracks[i].meta, "artist");
@@ -378,6 +382,7 @@ end:
         }
 
         cyanrip_log(ctx, 0, "%s\n", ctx->mb_submission_url);
+        crip_print_qrcode(ctx->mb_submission_url);
         if (ret)
             cyanrip_log(ctx, 0, "To continue add metadata via -a or -t, or ignore via -N!\n");
     }